// src/coolify-api/applications/types/application.ts
export interface CoolifyApplication {
  id: number;
  description?: string;
  repository_project_id?: number;
  uuid: string;
  name: string;
  fqdn?: string;
  config_hash: string;
  git_repository: string;
  git_branch: string;
  git_commit_sha: string;
  git_full_url?: string;
  docker_registry_image_name?: string;
  docker_registry_image_tag?: string;
  build_pack: string;
  static_image: string;
  install_command: string;
  build_command: string;
  start_command: string;
  ports_exposes: string;
  ports_mappings?: string;
  base_directory: string;
  publish_directory: string;
  health_check_enabled: boolean;
  health_check_path: string;
  health_check_port?: string;
  health_check_host?: string;
  health_check_method: string;
  health_check_return_code: number;
  health_check_scheme: string;
  health_check_response_text?: string;
  health_check_interval: number;
  health_check_timeout: number;
  health_check_retries: number;
  health_check_start_period: number;
  limits_memory: string;
  limits_memory_swap: string;
  limits_memory_swappiness: number;
  limits_memory_reservation: string;
  limits_cpus: string;
  limits_cpuset?: string;
  limits_cpu_shares: number;
  status: string;
  preview_url_template: string;
  destination_type: string;
  destination_id: number;
  source_id?: number;
  private_key_id?: number;
  environment_id: number;
  dockerfile?: string;
  dockerfile_location: string;
  custom_labels?: string;
  dockerfile_target_build?: string;
  manual_webhook_secret_github?: string;
  manual_webhook_secret_gitlab?: string;
  manual_webhook_secret_bitbucket?: string;
  manual_webhook_secret_gitea?: string;
  docker_compose_location: string;
  docker_compose?: string;
  docker_compose_raw?: string;
  docker_compose_domains?: string;
  docker_compose_custom_start_command?: string;
  docker_compose_custom_build_command?: string;
  swarm_replicas?: number;
  swarm_placement_constraints?: string;
  custom_docker_run_options?: string;
  post_deployment_command?: string;
  post_deployment_command_container?: string;
  pre_deployment_command?: string;
  pre_deployment_command_container?: string;
  watch_paths?: string;
  custom_healthcheck_found: boolean;
  redirect?: string;
  updated_at: string;
  deleted_at: string;
  compose_parsing_version: string;
}